Replacing old pandas code with new arrow-DuckDb code
