+++ cehr_bert.ini
@@ -4,10 +4,10 @@
 cdm_data_path = d:/GPM_CCAE
 
 # Maximum number of CPU cores to use:
-max_cores = 20
+max_cores = 5
 
 # Path where the patient sequence data will be written:
-output_path = d:/GPM_CCAE/patient_sequence
+output_path = d:/GPM_CCAE/patient_sequence_2
 
 
 [mapping]
